Add initial support for Death's Embrace.
Learned through example from other modules that due to the nature of how
Rawr operates and calculates data, it's easiest to simply average out
the damage increase of the talent, despite being conditional (based on
target health percentage).

In the case of Drain Soul, the talent is more potent due to the
interactions between the two, so an additional multiplier is instated
to account for the damage increase.