Adjust rotation for FDK when using Glyph of Disease without having Epidemic.
This impacts the Blood sub-spec variant of Frost. With a 15s disease
duration it's not possible to use death runes on Obliterate, without
dropping diseases, even when using Glyph of Disease. This adjusts the
rotation based on those two prerequisites. This change is supposed to
have no change whatsoever to the Unholy sub-spec, but it is currently
WIP and does.