Add damage proc effects to Enhancement Shamans.
For some reason they were not included in the model. This allows you to
see the damage dealt by trinkets that proc damage, such as Reign of the
Dead/Unliving. For some reason, the engineering rocket glove enchant
still doesn't show up despite this change.

Included in the comments is an original manual implementation, before I
discovered the proc helper. The manual implementation results in higher
DPS, though I am unsure of which is more accurate.