More gracefully handle missing images.
Instead of locking up the thread attempting to make network calls to an
address that no longer works, Rawr now uses a fallback icon (the famous
question mark icon) if the icon does not exist locally. It also
attempts to match the greyscale of the icon if applicable.

For now this is being treated as a temporary solution.