Clean up potentially interfering DLL files.
For some reason that I can't seem to work around, C# applications
look for a core DLL first in the base directory. For some stranger
reason, copies of Rawr that are distributed online containing a usable
cache, have an old version of the Base DLL located in the base
directory. Since users of this application are asked to overwrite the
files located in the old Rawr file, it is loading the old Base DLL and
it is interfering.

Solve this by renaming all of the existing DLLs located in the base
folder to an unusable file extension.