Fix fight length for Enhance damage proc item effects.
This went entirely undetected because the ICD of the trinket this system
was made for was lower than the fight duration even if it were in
seconds (ICD of 2s).

Trinkets with larger CDs such as 45s were accidentally seeing a
disproportionately large increase in DPS, as it was only counting the
worth of a trinket based on the first few seconds of an encounter and
not the additional, for example, ~40 seconds of downtime.

This fixes that. Notably Reign of the Unliving/Dead now has an even
larger gap between simmed DPS and DPS in practice, resulting in sims
further devaluing the trinket.